Move getASRouterLocalColumnFromJSON to fetchData.ts
